Added 'Recently Added'. Added backtracing on signal. Refactored logging
The times when episodes&seasons are added is now stored, and the "Recently Added" page lists  seasons which have been recently added, or have recently had episodes added to them.

If there's an unhandled exception, or a fatal signal, then a stack backtrace is now also logged.

The logging class has been refactored to be a bit cleaner, and more customisable, with the removal of some static code.